Working version with FFT Computation.

+++ ADC_FFT_Server/src/webserver.h
@@ -1,4 +1,4 @@
 #include <Arduino.h>
 
 String processor(const String& var);
-void webserverSetup(uint16_t*);
+void webserverSetup(double*);
